third_party/WebKit/LayoutTests/fast/css/variables/cssom-computed-style.html

<!doctype html>
<head><script src="../../js/resources/js-test-pre.js"></script></head>

<style>
#test-target {
    var-stylesheet: pass;
}
</style>

<body>
<div style="var-inherited: pass"><div id="test-target" style="var-inline: pass"></div></div>
</body>

<script>
description('This tests Javascript access to CSS variables via getComputedStyle.');

var computedStyle = getComputedStyle(document.querySelector("#test-target"));
var computedBodyStyle = getComputedStyle(document.querySelector("body"));

shouldBeEqualToString('computedStyle.var.toString()', '[object CSSVariablesMap]');
shouldBe('computedStyle.var === computedStyle.var', 'true');
shouldThrow('computedStyle.var.set("create", "test")', '"NoModificationAllowedError: Failed to set the \'create\' property on a computed \'CSSStyleDeclaration\': computed styles are read-only."');
shouldBe('computedStyle.var.size', '3');
shouldBe('computedBodyStyle.var.size', '0');
shouldBeEqualToString('computedStyle.var.get("stylesheet")', 'pass');
shouldBeEqualToString('computedStyle.var.get("inherited")', 'pass');
shouldBeEqualToString('computedStyle.var.get("inline")', 'pass');
shouldBeEqualToString('computedBodyStyle.var.get("non-existent")', '');

var forEachIterations = [];
computedStyle.var.forEach(function(value, name, map) {
    forEachIterations.push([value, name, map.toString()]);
});
forEachIterations.sort();
shouldBe('forEachIterations[0]', '["pass", "inherited", "[object CSSVariablesMap]"]');
shouldBe('forEachIterations[1]', '["pass", "inline", "[object CSSVariablesMap]"]');
shouldBe('forEachIterations[2]', '["pass", "stylesheet", "[object CSSVariablesMap]"]');
shouldBe('forEachIterations.length', '3');

var emptyVarForEachIterations = [];
computedBodyStyle.var.forEach(function(value, name, map) {
    emptyVarForEachIterations.push([value, name, map.toString()]);
});
shouldBe('emptyVarForEachIterations.length', '0');

shouldThrow('computedStyle.var.set("inline", "fail")', '"NoModificationAllowedError: Failed to set the \'inline\' property on a computed \'CSSStyleDeclaration\': computed styles are read-only."');
shouldThrow('computedStyle.var.clear()', '"NoModificationAllowedError: Failed to clear variables from a computed \'CSSStyleDeclaration\': computed styles are read-only."');
shouldBe('computedStyle.var.delete("inline")', 'false');
shouldBeEqualToString('computedStyle.var.get("inline")', 'pass');
</script>
<script src="../../js/resources/js-test-post.js"></script>
